Move security checks to users.js
